[LFC][IFC] Ceil the descent value when adjusting the layout bounds for an inline box https://bugs.webkit.org/show_bug.cgi?id=230557
Reviewed by Antti Koivisto.

Use ceil instead of floor to match legacy line layout positioning.

* layout/formattingContexts/inline/InlineDisplayContentBuilder.cpp: use ceilf everywhere
(WebCore::Layout::InlineDisplayContentBuilder::createRunsAndUpdateGeometryForLineContent):
* layout/formattingContexts/inline/InlineLineBoxBuilder.cpp:
(WebCore::Layout::LineBoxBuilder::adjustVerticalGeometryForInlineBoxWithFallbackFonts const):
(WebCore::Layout::LineBoxBuilder::setInitialVerticalGeometryForInlineBox const):


Canonical link: https://commits.webkit.org/241954@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@282823 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/layout/formattingContexts/inline/InlineDisplayContentBuilder.cpp
@@ -74,7 +74,7 @@ void InlineDisplayContentBuilder::createRunsAndUpdateGeometryForLineContent(cons
                 auto initialContaingBlockSize = RuntimeEnabledFeatures::sharedFeatures().layoutFormattingContextIntegrationEnabled()
                     ? formattingState.layoutState().viewportSize()
                     : formattingState.layoutState().geometryForBox(layoutBox.initialContainingBlock()).contentBox().size();
-                auto strokeOverflow = std::ceil(style.computedStrokeWidth(ceiledIntSize(initialContaingBlockSize)));
+                auto strokeOverflow = ceilf(style.computedStrokeWidth(ceiledIntSize(initialContaingBlockSize)));
                 auto inkOverflow = textRunRect;
                 inkOverflow.inflate(strokeOverflow);
                 auto letterSpacing = style.fontCascade().letterSpacing();
